\name{mailtools-package}
\alias{mailtools-package}
\alias{mailtools}
\docType{package}
\title{
E-Mail Utilties
}
\description{
E-Mail Utilties

}
\details{

  The package provides an R interface to several
  command line tools for sending emails.

}
\author{
Enrico Schumann

Maintainer: Enrico Schumann <es@enricoschumann.net>

}
\references{
to be added
}
\keyword{ package }
%% \seealso{
%% to be added
%% }
%% \examples{
%% ## to be added
%% }
