\name{msgID}
\alias{msgID}
\title{
  Message-ID
}
\description{
Create a unique message ID
}
\usage{
msgID()
}
\details{
  to be added
}
\value{
  a character
}
\references{
  Recommendations for generating Message IDs: \url{http://www.jwz.org/doc/mid.html}
}
\author{
Enrico Schumann
}

\seealso{
to be added
}
\examples{
##msgID()
}
